feat: flavours page now works

+++ src/main/java/hu/okrim/productratingapp/repository/FlavourRepository.java
@@ -1,8 +1,7 @@
 package hu.okrim.productratingapp.repository;
 
 import hu.okrim.productratingapp.entity.Flavour;
-import hu.okrim.productratingapp.entity.Person;
 import org.springframework.data.repository.CrudRepository;
 
-public interface FlavourRepository extends CrudRepository<Flavour, Long> {
+public interface FlavourRepository extends CrudRepository<Flavour, Integer> {
 }
